rename estimator module to regression
