+++ examples/plot_homoscedastic_1d_data.py
@@ -17,7 +17,7 @@ from sklearn.pipeline import Pipeline
 from sklearn.linear_model import LinearRegression
 from matplotlib import pyplot as plt
 
-from mapie.estimators import MapieRegressor
+from mapie.regression import MapieRegressor
 
 
 def f(x: np.ndarray) -> np.ndarray:
